Fixed #308, #309, #313. No more "Wrong parameter in the URL". Anti-XSS now acts much more polite. It doesn't block the page completely but adds a message to $error_string. Most forms check that string and display the error message. It also saves you from refilling the form again.
The user_birthdate was changed to DATE type to process it with SQL queries as well as PHP code. For example, now you can select users who have birthday this date but before 1970 with a simple SQL query. As a side effect, most scripts use the old timestamp format, so you need to use sed_date2stamp() and sed_stamp2date() functions to convert DATE to TIMESTAMP and vice versa.

There is a problem about converting existing databases to use DATE, MySQL can't handle that alone. I've created a patch script for that: sql/patch-0.0.6.php. Run it if you are updating to 0.0.6.

These are quite big changes. So please test.