First step for #395. Created file for resource strings, moved icons from common and functions, moved some static code from PFS.
As for the PFS, I have also applied simple string optimizations, it may make sense if applied everywhere. Sorry if I have broken something, but it seems the PFS changes are not complete yet. Some HTML needs to be moved to TPLs too. And some sed_url() stuff is missing.